moved 'get_project_database' and 'get_project_connection' to mongo
